Struct tensorflux::Session

pub struct Session { /* fields omitted */ }

A session.

Methods

impl Session

fn new(options: &Options) -> Result<Self>

Create a session.

fn extend(&mut self, definition: &Buffer) -> Result<()>

Extend the graph using a protocol buffer.

The schema of the definition protocol buffer is called GraphDef, and it can be found in TensorFlow’s repository. An example of creating a graph definition is given in the main description of this package.

fn run(&mut self, inputs: &[Input], outputs: &mut [Output], targets: &[Target], options: Option<&Buffer>, metadata: Option<&mut Buffer>) -> Result<()>

Run the graph.

The schemas of the options and metadata protocol buffers are called RunOptions and RunMetadata, respectively, and they can be found in TensorFlow’s repository.